sequence_list : a t list -> a list t
I need to recursively shrink a list of values (I want to enumerate
lists of all shrunk values) and I don't see how to do it without this
function.

Haskell typically exposes both (List specializing the traversable
datatype here)

    sequence : Applicative f => f (List a) -> f (List a)
    traverse : Applicative f => (a -> f b) -> List a -> f (List b)

For now I propose only `sequence` to keep the API simpler, and also
because it is a natural (homogeneous) extension of pair, triple, quad.